requirements.txt
Django==2.1.7
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-formtools==2.1
djangorestframework==3.9.2
httplib2==0.9.2
mysqlclient==1.3.14
pytz==2018.9

#Requirements for dashboard
ansi2html==1.5.2
ansible==2.7.8
asn1crypto==0.24.0
autopep8==1.4.3
bcrypt==3.1.6
cffi==1.12.2
cryptography==2.6.1
Django==2.1.7
Jinja2==2.10
MarkupSafe==1.1.1
obs-websocket-py==0.4
paramiko==2.4.2
Pillow==5.4.1
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
PyNaCl==1.3.0
pytz==2018.9
PyYAML==3.13
six==1.12.0
websocket-client==0.55.0